Added photo delete confirmation [#2124 state:qa]

+++ public/javascripts/zz/photo.js
@@ -282,17 +282,19 @@
         //delete
         delete_photo:  function(){
             var self = this;
-            if(self.options.onDelete()){
-                self.captionElement.hide();
-                self.deleteButtonElement.hide();
-                self.borderElement.hide("scale", {}, 300, function(){
-                    self.element.animate({width:0},500, function(){
-                        self.element.remove();
-                        if(self.photoGrid){
-                            self.photoGrid.resetLayout();
-                        }
-                    })
-                });
+            if(confirm("Are you sure you want to delete this photo?")){
+                if(self.options.onDelete()){
+                    self.captionElement.hide();
+                    self.deleteButtonElement.hide();
+                    self.borderElement.hide("scale", {}, 300, function(){
+                        self.element.animate({width:0},500, function(){
+                            self.element.remove();
+                            if(self.photoGrid){
+                                self.photoGrid.resetLayout();
+                            }
+                        })
+                    });
+                }
             }
         },
